2_openai/deep_research/requirements.txt
gradio==5.49.1
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
aiofiles>=23.2.1
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
websockets>=11.0.3
sendgrid>=6.11.0